hot fix

+++ story-service/src/main/java/com/onlinestories/story_service/Service/StoryService.java
@@ -195,8 +195,10 @@ public class StoryService {
                 genres.addAll(newGenres);
             }
             if (img != null && !img.isEmpty()) {
-                String message = mediaClient.deleteFile(story.getImg(), "image");
-                log.info("Deleted old image for story {}: {}", story.getStoryId(), message);
+                if (story.getImg() != null) {
+                    String message = mediaClient.deleteFile(story.getImg(), "image");
+                    log.info("Deleted old image for story {}: {}", story.getStoryId(), message);
+                }
                 String imgUrl = mediaClient.uploadFile(img, "cover-img");
                 story.setImg(imgUrl);
             }
